refactor: 뒤로가기 icon 추가 및 icon 디렉토리 구조 변경
